Avoid uci error and wget output in logfile every minute
checkuplink is writing 5 lines of log output every minute. Due to the limited logread buffer this quickly removes older log entries, which might be important for error analysis. This change removes the uci error and the wget output, so checkuplink will only generate log output, in case a reconnection is needed.