Also check BATMAN connectivity, after wireguard
Sometimes it happens that the wireguard tunnel is up and running, but BATMAN connectivity is lost. In this case checkuplink does not detect the node is actually disconnected and yanic reporting also keeps working, so it does appear online in the map, but client network for users is not working. This change adds an additional check, whether the current BATMAN gateway is pingable, so BATMAN networking is working as expected.