fix invalid HTTPS check
The old code ran wget and checked only if exit code was '1' and then used
HTTP as a fallback.

wget returns several different exit code:
https://www.gnu.org/software/wget/manual/html_node/Exit-Status.html

In case for example an "SSL verification failure." occured (expired certs)
the fallback did not trigger and it tried but failed to establish an HTTPS
connection.