Add ffbs-debugbathosts
This package provides means to map MAC addresses to names for `batctl`.

This is done with two components:

* `/etc/bat-hosts` is a mapping file consumed by `batctl`.
  It is delivered empty and only contains help for the user.
  It is mostly intended as a reminder that this functions exists at all.
* `update-bat-hosts` is a lua-script that loads a `nodes.json` from a given URL
  and populates `/etc/bat-hosts` with the nodes found.

The script `update-bat-hosts` is only intended to be run by an
experienced user by hand to ease debugging. The result is directly written
into to `/etc/` and thus, usually, into flash memory. The
number of write/erase cycles caused by this should be negligible.