ffbs-mesh-vpn-parker: depend on simple-tc
This is a preparation to later add traffic shaping to nodeconfig.
For traffic shaping `simple-tc` is needed. So let's depend on it.

While we are on it: Remove non-functional code from the VPN-provider.
We do not know which interfaces we will have during running-mode, so we
can not configure traffic limits for those on reconfigure.
(It has to be done when dynamically setting up the
Wireguard-interfaces.)