ffbs-mesh-vpn-pakrer: Fix nodeconfig.lua exiting due to a nil-value
Since Bob Idle's typo fix in b37699a8dafedb60026746b7f4932b961f1aa2bd
`nodeconfig.lua` would exit with a lua stacktrace when executed:
> nodeconfig.lua: Starting up
> nodeconfig.lua: Updating MTU on wg-interface wg_c1 to 1375
> nodeconfig.lua: Updating MTU on wg-interface wg_c3 to 1375
> lua: /usr/share/lua/nodeconfig.lua:110: attempt to concatenate field 'keepalive' (a nil value)
> stack traceback:
>         /usr/share/lua/nodeconfig.lua:110: in function 'apply_wg'
>         /usr/share/lua/nodeconfig.lua:176: in main chunk
>         [C]: ?

Why this happens is an interesting story that I want to share:

During review of the commit in question I grep'ped for the fixed name
and did not find any match. I assumed that we never used the value and
that the fix would thus not make any difference.
But I was so wrong.

With the typo fix `cur_conf.keepalive` (that's just another name for
`peer.persistent_keepalive`) in `nodeconfig.lua` would now be
a value different from `nil`.
And that is a problem, because `target.keepalive` does simply not exist.
That means that the values will now be different and we will try to
access `target.keepalive`. And this is what actually fails.

Why does `target.keepalive` not exist and why didn't we realize in all
these years? Well:

* The typo in `util.lua` was present since 2018
  (https://gitli.stratum0.org/ffbs/ffbs-packages/-/commit/6813cefc056a85f200405681fb1fc2bdca720ea3 ).
  That is actually the first commit for this file.
  This means `cur_conf.keepalive` was `nil` since the beginning of time.
* Also in this commit: We never read the actual keepalive value from the
  configuration, but used `transfer_tx` as a `keepalive`.
  This wasn't really a problem, since we never accessed the value
  because of the typo anyway.
* At some point in 2020
  (https://gitli.stratum0.org/ffbs/ffbs-packages/-/commit/6f1fb8f2a1b4bf4c100e1b90388b3c65f722c977 )
  we reworked `nodeconfig.lua` and introduced the non-existent
  `target.keepalive`. But this seemingly worked because of the typo.
* In the end `nodeconfig.lua` was never able to update the keepalive
  timeout of an interface. But I assume we never tried to change this
  value anyway.

This commit fixes all these errors. `nodeconfig.lua` should now be able
to set the correct keepalive timeout, if it was ever to be reconfigured.