ffbs-mesh-vpn-pakrer: nodeconfig.lua: Apply traffic limits
This change applies traffic limits to all Wireguard interfaces created.
The traffic limits set by the user in config mode are used.

This is not a final solution, since users can exceed the limits
configured (see below).
But it is at least better than not implement traffic limiting at all
(while still presenting the option on the web interface)!

The traffic limit is applied to all Wiregard interfaces created.
This means that a single flow (e.g. TCP connection) will very likely be
shaped as expected - assuming that the route (which tunnel it takes)
stays constant most of the time.

For multiple flows (that will likely take different routes / tunnels)
however, limiting will likely not work as expected.
This is because multiple flows can take multiple tunnels and thus use up
multiples of the configured bandwidth limit.