ffbs-mesh-vpn-parker: noderoute.sh: Fix logger shaddowing returncode
Using the pipe of stdout into the logger we loose the return code of
`noderoute.lua` here.
The `noderoute.lua` actually already logs into it's own logger, so there
is nothing to gain here from capturing stdout. So let's drop that.

This way we can actually log, if the rc!=0.