ffbs-mesh-vpn-parker: nodeconfig.sh: Report tc limits to config-server
This change is a preparation for traffic shaping on the
Wiregard-Endpoints (aka. concentrators).

Usually Gluon nodes shape traffic in egress and ingress on the node
itself. Especially for ingress this is problematic, since the node can
only drop packets once it has received to many.
(For egress it's no problem to shape on the node.)

Once the config-service and in extend the concentrators know about the
configured traffic limit we _could_ shape ingress on the concentrators
only.
(In this case we would likely also drop the ingress limit applied in
nodeconfig.lua, if the config-response tells us so.)